pyproject.toml
[project]
name = "amplifier-bundle-plugin-compat"
version = "0.1.0"
description = "Claude Code plugin compatibility for Amplifier"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Amplifier Team" }
]
keywords = ["amplifier", "claude-code", "plugins", "compatibility"]

dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "gitpython>=3.1",
]

[project.scripts]
amplifier-plugins = "amplifier_plugin_compat.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["amplifier_plugin_compat"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "UP"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
